Zero-configuration VisualEditor + PHP for MediaWiki release 1.37
This copies the Parsoid extension code into includes/VEParsoid
to allow a "one extension, zero-configuration" install of
VisualEditor for MW's release. The Parsoid code has been
re-namespaced (`VEParsoid` instead of `MWParsoid`) to avoid
autoloader conflicts if you actually install Parsoid as an
extension (as we do in Wikimedia production). Similarly, we
arrange that the ServiceWiring and RestRoutes configurations
are skipped unless running in zeroconf mode, to avoid
conflicts with the Parsoid extension.

This import matches the Parsoid commit tagged v0.14.0-a18
(cfe9391c047b85feaf72f745f9b6ec0d34368d96).

It is not-quite-a-cherry-pick from VE commit
09aed3b206dad3f5361fafe6d26650d840acacb6.

Bug: T248343
Bug: T261220
Bug: T291437
Change-Id: Iac8249bc80dcc0fac3c9cf621e7fb1502e09872e